% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_model.R
\name{gdp_model}
\alias{gdp_model}
\title{Fit a linear model predicting total plastic waste}
\usage{
gdp_model(explanatory_var = "gdp_per_capita_nominal")
}
\arguments{
\item{explanatory_var}{A string giving the name of the explanatory variable to use in the model.
Must be a column in the data returned by load_data}
}
\value{
A linear model object of class lm
}
\description{
Fit a linear model predicting total plastic waste
}
