% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_plastics.R
\name{gdp_plastics}
\alias{gdp_plastics}
\title{GDP Plastics Summary Function}
\usage{
gdp_plastics(input_country)
}
\arguments{
\item{input_country}{Country that is getting the totals calculated for}
}
\value{
A tibble containing the plastic total and the average GDP per capita for the input country
}
\description{
Calculates the total plastic and the average GDP per capita for each country
}
