% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_plot.R
\name{gdp_plot}
\alias{gdp_plot}
\title{Creates a scatter plot}
\usage{
gdp_plot(cutoff1 = 5000, cutoff2 = 20000)
}
\arguments{
\item{cutoff1}{the lowest cutoff for differentiating Low and Mid GDP}

\item{cutoff2}{the highest cutoff for differentiating Mid and High GDP}
}
\value{
A scatterplot showing the relationship between average GDP per capita and plastic waste
coloring by GDP category and labelling the countries
}
\description{
Plots the relationship between average GDP per capita and plastic waste for each country
}
