% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loads in the plastic data}
\usage{
load_data()
}
\arguments{
\item{csv_data}{The csv data set that is getting read into the global environment}
}
\value{
The csv data set as a data frame
}
\description{
Loads in the plastic data
}
