% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plastics_long.R
\name{make_plastics_long}
\alias{make_plastics_long}
\title{Create the plastics_long dataset}
\usage{
make_plastics_long()
}
\value{
A long formatted version of plastics top,
with the total plastic, plastic type, country, year, and company being returned
}
\description{
Turns the plastics_top data into a long format so that a Sankey diagram can be made
}
