% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_plastics_table.R
\name{gdp_plastics_table}
\alias{gdp_plastics_table}
\title{GDP and Plastics Summary Table}
\usage{
gdp_plastics_table(top_x_countries = 10)
}
\arguments{
\item{top_x_countries}{the number of top countries the user wants to display}
}
\value{
A data frame containing the top countries by total plastic waste,
along with their total plastic waste and average GDP per capita.#'
}
\description{
Creates a summary table of the countries with the highest plastic waste,
displaying each country's total plastic waste and average GDP per capita.
}
\examples{
gdp_plastics_table(5)
}
