% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_plot.R
\name{sankey_plot}
\alias{sankey_plot}
\title{Creates a Sankey plot}
\usage{
sankey_plot(top_x_countries = 2, top_x_companies = 5)
}
\arguments{
\item{top_x_countries}{the number of top countries the user wants to display}

\item{top_x_companies}{the number of top companies the user wants to display}
}
\value{
A Sankey diagram showing the relationship between country and company for different types of
plastic waste
}
\description{
Plots the relationship between country and company when it comes to plastic waste
}
